models refactor start
