+++ packages/core/src/chat/MODELS.ts
